+++ CHANGELOG.md
@@ -21,6 +21,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   default. Previously some functions that took `O_*` flags would only set
   `O_CLOEXEC` if the user explicitly requested it, but `O_CLOEXEC` is easy to
   unset on file descriptors and having it enabled is a more sane default.
+- The C API values of the `pathrs_proc_base_t` enum (`PATHRS_PROC_BASE_*`) have
+  different values, in order to support `ProcfsBase::ProcPid` passing from C
+  callers. Any binaries compiled with the old headers will need to be
+  recompiled to avoid spurious behaviour.
+  - This required a breaking change in the Go bindings for libpathrs.
+    `ProcfsBase` is now an opaque `struct` type rather than a simple `int`
+    wrapper -- this was necessary in order to add support for
+    `ProcfsBase::ProcPid` in the form of the `ProcBasePid` helper function.
 
 ### Added ###
 - python bindings: add `Root.creat_raw` to create a new file and wrap it in a
@@ -43,6 +51,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   logic of libpathrs is race-safe against various attacks. This is no big
   surprise, given libpathrs's design, but we now have more extensive tests than
   `github.com/cyphar/filepath-securejoin`.
+- procfs: added `ProcfsBase::ProcPid(n)` which is just shorthand when operating
+  on a operating on a different process. This is also now supported by the C
+  API (by just passing the `pid_t` instead of a special `pathrs_proc_base_t`
+  value).
 
 ### Changes ###
 - procfs: our internal `GLOBAL_PROCFS_HANDLE` has been removed entirely, and